# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Eye of Providence API
  version: "1.0.0"
  description: |
    Public-facing endpoints для интеграций (BI, custom dashboards, CI batchers,
    third-party clients).

    Покрытие: auth + me + public read API + ingest. Внутренние admin endpoints
    не документированы — стабильность не гарантирована. Контракт для public
    API: backward-compatible до v2.

    Auth:
      - JWT (Bearer eyJ...) — выдаётся /v1/auth/login или /v1/auth/register.
      - API token (Bearer eop_<48 hex>) — long-lived, scope-based.
        Создание/revoke через /v1/me/tokens.

    Scopes:
      - read         — только чтение (events, summary, languages, trend, insights)
      - write:ingest — POST /v1/ingest
      - admin        — full account access

  contact:
    name: Eye of Providence
    url: https://eop.rysdavletov.org
servers:
  - url: https://eop.rysdavletov.org/api
    description: Production
  - url: http://localhost:8080
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Регистрация
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, display_name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                display_name: { type: string, maxLength: 64 }
                invite_code: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login по email+password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Запрос на password reset (email)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { email: { type: string, format: email } }
      responses:
        "200":
          description: |
            Всегда 200, чтобы не палить наличие email'а в системе. Если email
            существует — отправлено письмо с reset-токеном.
  /v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Применить reset-токен
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200": { description: OK }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /v1/me/insights:
    get:
      tags: [Me]
      summary: Narrative-инсайты (key + i18n vars)
      parameters:
        - in: query
          name: tz
          schema: { type: string, default: "UTC" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  insights:
                    type: array
                    items: { $ref: "#/components/schemas/Insight" }
  /v1/me/tokens:
    get:
      tags: [Me]
      summary: Список API tokens (без plaintext)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    type: array
                    items: { $ref: "#/components/schemas/APIToken" }
    post:
      tags: [Me]
      summary: Создать API token (plaintext возвращается ровно раз)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 64 }
                scope:
                  type: string
                  enum: [read, "write:ingest", admin]
                  default: read
                ttl_days: { type: integer, minimum: 0, maximum: 365, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: "Plaintext, ровно раз" }
                  metadata: { $ref: "#/components/schemas/APIToken" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT or eop_<token>"
  schemas:
    AuthResponse:
      type: object
      properties:
        token: { type: string, description: "JWT" }
        user_id: { type: string, format: uuid }
        display_name: { type: string }
        team_id: { type: string, format: uuid, nullable: true }
...
    Insight:
      type: object
      required: [key]
      properties:
        key:
          type: string
          description: |
            i18n key. Possible: ai_trend_up, ai_trend_down, ai_trend_flat,
            ai_trend_started, ai_ratio, top_lang, productive_day, total_activity.
        vars:
          type: object
          additionalProperties: true
    APIToken:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        scope: { type: string, enum: [read, "write:ingest", admin] }
        prefix: { type: string, description: "First 8 chars (eop_xxx)" }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, nullable: true }
        last_used_at: { type: string, format: date-time, nullable: true }
...
    Error:
      type: object
      properties:
        error: { type: string }
        request_id: { type: string }
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json: { schema: { $ref: "#/components/schemas/Error" } }
    Unauthorized:
      description: Unauthorized
      content:
        application/json: { schema: { $ref: "#/components/schemas/Error" } }
...
    Conflict:
      description: Conflict
      content:
        application/json: { schema: { $ref: "#/components/schemas/Error" } }